back/requirements.txt
django-cors-headers==3.14.0
djangorestframework
python-dotenv
entrypoints==0.4
executing==1.2.0
flatbuffers==23.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.0.4
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
openai
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.0
six==1.16.0
smmap==5.0.0
sqlparse 
stack-data==0.6.2
streamlit==1.20.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions 
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.15.0